fix: move test out folder
