+++ backend/internal/search/handler.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/amrllkmn/thoth/backend/internal/utils"
 	"github.com/gin-gonic/gin"
+	"github.com/labstack/gommon/log"
 )
 
 type SQLiteSearchHandler struct {
@@ -14,6 +15,7 @@ type SQLiteSearchHandler struct {
 func (h *SQLiteSearchHandler) FindAll(c *gin.Context) {
 	books, err := h.searchService.FindAll()
 	if err != nil {
+		log.Error(err)
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"message": "Something went wrong",
 		})
@@ -30,6 +32,7 @@ func (h *SQLiteSearchHandler) FindByQuery(c *gin.Context) {
 	query := c.Query("query")
 	books, err := h.searchService.FindByQuery(query)
 	if err != nil {
+		log.Error(err)
 		c.JSON(http.StatusInternalServerError, gin.H{
 			"message": "Something went wrong",
 		})
